backend/requirements.txt
django==3.2.16
django-filter==23.4
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
PyJWT==2.1.0
requests==2.26.0
djoser==2.1.0
flake8==5.0.4
Pillow==11.0.0
drf-extra-fields==3.7.0